+++ CHANGELOG.md
@@ -1,1 +1,5 @@
+# Changelog
 
+## 0.4.1
+- use store/state for language
+- fix handling of unknown language.